Revert workaround for indexing invalid concepts
This reverts the changes commit 6766c04a22b3bc71848ad1c9952297411a432cd1

Due to NTRF import erroneously allowing import of concepts with identifiers
beginning with a numeric character, IndexService attempted to index the
terminology itself as a concept, causing indexing to fail.

Now that the NTRF import should no longer allow concepts with invalid
identifiers, and the existing concepts in the database should be fixed,
this workaround can be removed.